+++ react-client/src/components/PlatformAnalysis.tsx
