+++ react-client/src/components/PlatformAnalysisBox.tsx
