+++ react-client/src/components/legacy/DefaultSearchBar.tsx
