+++ react-client/src/components/legacy/DraggableBox.tsx
