+++ react-client/src/components/legacy/DraggableBoxWrapper.tsx
