+++ react-client/src/components/legacy/DynamicResizableComponent.tsx
